scripts/painting-size-investigation/requirements.txt
requests
lxml
Pillow
playwright
